# packages/contracts/openapi/v1.yaml
...
openapi: 3.1.0
info:
  description: |-
    Version 1 of the Chaya 02 HTTP API. Breaking changes require a new /api/v2 document.

    Generated from the backend's controllers; packages/contracts/openapi/v1.yaml is the committed output and OpenApiContractTest fails when it drifts.

    Roles (x-chaya-roles) under /api/v1/venues/{venueId}/ are the roles the caller holds at that venue. A resource outside the caller's organization or venue answers 404, never 403.

    Errors are RFC 9457 problems with a stable machine-readable `code`.
  title: Chaya 02 API
  version: v1
servers:
- description: The API's own origin
  url: /
paths:
...
  /api/v1/venues/{venueId}/floors/{floorId}/rescan:
    post:
      description: See docs/incremental-rescan.md.
      operationId: rescanInitiate
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: floorId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescanRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RescanInitiated"
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "REGION_TOO_SMALL, REGION_TOO_LARGE, INVALID_JSON"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "VERSION_NOT_FINALIZED, VERSION_WRONG_FLOOR, NOT_CALIBRATED"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Start a region re-scan of the floor from a FINALIZED source version
      tags:
      - rescan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
  /api/v1/venues/{venueId}/floors/{floorId}/scan-versions:
    get:
      operationId: rescanListVersions
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: floorId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanVersionView"
          description: OK
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: "The floor's scan versions, newest first"
      tags:
      - rescan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
      - VIEWER
  /api/v1/venues/{venueId}/floors/{floorId}/scan-versions/finalize-current:
    post:
      operationId: rescanFinalizeCurrent
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: floorId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanVersionView"
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Malformed request: an unparseable parameter or body, or a\
            \ failed validation (code BAD_REQUEST or an operation-specific code)"
        "401":
          description: "Missing, invalid or expired credentials (empty body, WWW-Authenticate\
            \ header)"
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "The caller's roles do not allow this operation (empty body),\
            \ or an operation-specific refusal (problem with a code)"
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Not found, or not visible to the caller's organization or\
            \ venue (code NOT_FOUND)"
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "VERSION_INCOMPLETE, VERSION_INCONSISTENT, NOT_CALIBRATED,\
            \ or nothing to finalize"
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Rate limit exceeded (code RATE_LIMITED); see the Retry-After
            header
        "503":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "A dependency is unavailable (code DATABASE_UNAVAILABLE, STORAGE_UNAVAILABLE\
            \ or AUTHENTICATION_UNAVAILABLE); the request was not completed"
      security:
      - bearerAuth: []
      summary: Bootstrap version 1 from the floor's latest successful full-venue reconstruction
      tags:
      - rescan-controller
      x-chaya-roles:
      - ADMIN
      - VENUE_MANAGER
      - OPERATOR
...
components:
  schemas:
...
    PinnedArtifact:
      type: object
      properties:
        artifactId:
          type: string
          format: uuid
        kind:
          type: string
        ownerVersionId:
          type: string
          format: uuid
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        code:
          type: string
          description: "Stable machine-readable error code, e.g. NOT_FOUND, BAD_REQUEST,\
            \ VERSION_INCOMPLETE. Absent on Spring's own request-parsing and validation\
            \ problems."
        detail:
          type: string
        instance:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
          example: about:blank
      required:
      - status
...
    RegionGeometry:
      type: object
      properties:
        points:
          type: array
          items:
            type: array
            items:
              type: number
              format: double
            maxItems: 2
            minItems: 2
          maxItems: 2147483647
          minItems: 3
      required:
      - points
...
    RescanInitiated:
      type: object
      properties:
        captureId:
          type: string
          format: uuid
        navigationRebuildRequired:
          type: boolean
        parentVersionId:
          type: string
          format: uuid
        regionAreaSquareMeters:
          type: number
          format: double
        scanVersionId:
          type: string
          format: uuid
    RescanRequest:
      type: object
      properties:
        device:
          type: object
          additionalProperties: {}
        parentVersionId:
          type: string
          format: uuid
        region:
          $ref: "#/components/schemas/RegionGeometry"
      required:
      - parentVersionId
      - region
...
    ScanVersionView:
      type: object
      properties:
        alignmentConfidence:
          type: number
          format: double
        alignmentMethod:
          type: string
        alignmentReport:
          type: object
          additionalProperties: {}
        alignmentResidualM:
          type: number
          format: double
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/PinnedArtifact"
        changedArtifactKinds:
          type: array
          items:
            type: string
        coordinateFrameId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        current:
          type: boolean
        finalizedAt:
          type: string
          format: date-time
        floorId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        parentVersionId:
          type: string
          format: uuid
        pipelineRunId:
          type: string
          format: uuid
        processingConfig:
          type: object
          additionalProperties: {}
        regionGeometry:
          type: object
          additionalProperties: {}
        rejectedAt:
          type: string
          format: date-time
        scanId:
          type: string
          format: uuid
        spliceReport:
          type: object
          additionalProperties: {}
        status:
          type: string
        versionNumber:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Keycloak access token for the chaya-api audience. Organization,\
        \ roles and venue memberships come from its claims."
      scheme: bearer
      type: http
    viewerToken:
      description: "Short-lived public-viewer token from POST /api/v1/public/viewer-token,\
        \ bound to one venue; acts as PUBLIC_VIEWER. Never together with a bearer\
        \ token (400)."
      in: header
      name: X-Chaya-Viewer-Token
      type: apiKey